update vAjusteStock, vSeleccionProductoAjuste, mAjusteStock, AjusteStockDAO y cAjusteStock

+++ src/main/java/modelo/AjusteStockDAO.java
@@ -6,6 +6,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class AjusteStockDAO {
+
     private Connection conexion;
 
     public AjusteStockDAO() throws SQLException {
@@ -14,23 +15,23 @@ public class AjusteStockDAO {
 
     // Buscar producto por código de barras con stock actual
     public Object[] buscarProductoConStock(String codBarra) throws SQLException {
-        String sql = "SELECT pc.id_producto, pc.nombre, pd.cod_barra, pd.descripcion, " +
-                    "COALESCE(s.cantidad_disponible, 0) as stock " +
-                    "FROM productos_cabecera pc " +
-                    "INNER JOIN productos_detalle pd ON pc.id_producto = pd.id_producto " +
-                    "LEFT JOIN stock s ON pd.id_producto = s.id_producto AND pd.cod_barra = s.cod_barra " +
-                    "WHERE pd.cod_barra = ? AND pc.estado = 1 AND pd.estado = 1";
+        String sql = "SELECT pc.id_producto, pc.nombre, pd.cod_barra, pd.descripcion, "
+                + "COALESCE(s.cantidad_disponible, 0) as stock "
+                + "FROM productos_cabecera pc "
+                + "INNER JOIN productos_detalle pd ON pc.id_producto = pd.id_producto "
+                + "LEFT JOIN stock s ON pd.id_producto = s.id_producto AND pd.cod_barra = s.cod_barra "
+                + "WHERE pd.cod_barra = ? AND pc.estado = 1 AND pd.estado = 1";
 
         try (PreparedStatement ps = conexion.prepareStatement(sql)) {
             ps.setString(1, codBarra);
             try (ResultSet rs = ps.executeQuery()) {
                 if (rs.next()) {
                     return new Object[]{
-                        rs.getInt("id_producto"),      // 0
-                        rs.getString("nombre"),        // 1
-                        rs.getString("cod_barra"),     // 2
-                        rs.getString("descripcion"),   // 3
-                        rs.getDouble("stock")          // 4
+                        rs.getInt("id_producto"), // 0
+                        rs.getString("nombre"), // 1
+                        rs.getString("cod_barra"), // 2
+                        rs.getString("descripcion"), // 3
+                        rs.getInt("stock")
                     };
                 }
             }
@@ -40,8 +41,8 @@ public class AjusteStockDAO {
 
     // Insertar cabecera de ajuste
     public int insertarAjusteCabecera(mAjusteStock ajuste) throws SQLException {
-        String sql = "INSERT INTO ajustes_stock_cabecera (fecha, observaciones, usuario_id, aprobado, estado) " +
-                    "VALUES (?, ?, ?, ?, ?)";
+        String sql = "INSERT INTO ajustes_stock_cabecera (fecha, observaciones, usuario_id, aprobado, estado) "
+                + "VALUES (?, ?, ?, ?, ?)";
 
         try (PreparedStatement ps = conexion.prepareStatement(sql, Statement.RETURN_GENERATED_KEYS)) {
             ps.setTimestamp(1, new Timestamp(ajuste.getFecha().getTime()));
@@ -63,15 +64,15 @@ public class AjusteStockDAO {
 
     // Insertar detalle de ajuste
     public void insertarAjusteDetalle(int idAjuste, mAjusteStock.DetalleAjuste detalle) throws SQLException {
-        String sql = "INSERT INTO ajustes_stock_detalle (id_ajuste, id_producto, cod_barra, " +
-                    "cantidad_sistema, cantidad_ajuste, observaciones) VALUES (?, ?, ?, ?, ?, ?)";
+        String sql = "INSERT INTO ajustes_stock_detalle (id_ajuste, id_producto, cod_barra, "
+                + "cantidad_sistema, cantidad_ajuste, observaciones) VALUES (?, ?, ?, ?, ?, ?)";
 
         try (PreparedStatement ps = conexion.prepareStatement(sql)) {
             ps.setInt(1, idAjuste);
             ps.setInt(2, detalle.getIdProducto());
             ps.setString(3, detalle.getCodBarra());
-            ps.setDouble(4, detalle.getCantidadSistema());
-            ps.setDouble(5, detalle.getCantidadAjuste());
+            ps.setInt(4, detalle.getCantidadSistema());
+            ps.setInt(5, detalle.getCantidadAjuste());
             ps.setString(6, detalle.getObservaciones());
 
             ps.executeUpdate();
@@ -80,8 +81,8 @@ public class AjusteStockDAO {
 
     // Buscar ajuste por ID
     public mAjusteStock buscarAjustePorId(int idAjuste) throws SQLException {
-        String sqlCabecera = "SELECT id_ajuste, fecha, observaciones, usuario_id, aprobado, estado " +
-                           "FROM ajustes_stock_cabecera WHERE id_ajuste = ?";
+        String sqlCabecera = "SELECT id_ajuste, fecha, observaciones, usuario_id, aprobado, estado "
+                + "FROM ajustes_stock_cabecera WHERE id_ajuste = ?";
 
         try (PreparedStatement ps = conexion.prepareStatement(sqlCabecera)) {
             ps.setInt(1, idAjuste);
@@ -107,13 +108,13 @@ public class AjusteStockDAO {
     // Cargar detalles de un ajuste
     private List<mAjusteStock.DetalleAjuste> cargarDetallesAjuste(int idAjuste) throws SQLException {
         List<mAjusteStock.DetalleAjuste> detalles = new ArrayList<>();
-        
-        String sql = "SELECT asd.id_detalle, asd.id_producto, asd.cod_barra, asd.cantidad_sistema, " +
-                    "asd.cantidad_ajuste, asd.observaciones, pc.nombre, pd.descripcion " +
-                    "FROM ajustes_stock_detalle asd " +
-                    "INNER JOIN productos_cabecera pc ON asd.id_producto = pc.id_producto " +
-                    "INNER JOIN productos_detalle pd ON asd.id_producto = pd.id_producto AND asd.cod_barra = pd.cod_barra " +
-                    "WHERE asd.id_ajuste = ? ORDER BY asd.id_detalle";
+
+        String sql = "SELECT asd.id_detalle, asd.id_producto, asd.cod_barra, asd.cantidad_sistema, "
+                + "asd.cantidad_ajuste, asd.observaciones, pc.nombre, pd.descripcion "
+                + "FROM ajustes_stock_detalle asd "
+                + "INNER JOIN productos_cabecera pc ON asd.id_producto = pc.id_producto "
+                + "INNER JOIN productos_detalle pd ON asd.id_producto = pd.id_producto AND asd.cod_barra = pd.cod_barra "
+                + "WHERE asd.id_ajuste = ? ORDER BY asd.id_detalle";
 
         try (PreparedStatement ps = conexion.prepareStatement(sql)) {
             ps.setInt(1, idAjuste);
@@ -124,8 +125,8 @@ public class AjusteStockDAO {
                     detalle.setIdAjuste(idAjuste);
                     detalle.setIdProducto(rs.getInt("id_producto"));
                     detalle.setCodBarra(rs.getString("cod_barra"));
-                    detalle.setCantidadSistema(rs.getDouble("cantidad_sistema"));
-                    detalle.setCantidadAjuste(rs.getDouble("cantidad_ajuste"));
+                    detalle.setCantidadSistema(rs.getInt("cantidad_sistema"));
+                    detalle.setCantidadAjuste(rs.getInt("cantidad_ajuste"));
                     detalle.setObservaciones(rs.getString("observaciones"));
                     detalle.setNombreProducto(rs.getString("nombre"));
                     detalle.setDescripcionProducto(rs.getString("descripcion"));
@@ -140,9 +141,8 @@ public class AjusteStockDAO {
     // Obtener primer ajuste
     public mAjusteStock obtenerPrimerAjuste() throws SQLException {
         String sql = "SELECT id_ajuste FROM ajustes_stock_cabecera ORDER BY id_ajuste ASC LIMIT 1";
-        
-        try (PreparedStatement ps = conexion.prepareStatement(sql);
-             ResultSet rs = ps.executeQuery()) {
+
+        try (PreparedStatement ps = conexion.prepareStatement(sql); ResultSet rs = ps.executeQuery()) {
             if (rs.next()) {
                 return buscarAjustePorId(rs.getInt("id_ajuste"));
             }
@@ -153,9 +153,8 @@ public class AjusteStockDAO {
     // Obtener último ajuste
     public mAjusteStock obtenerUltimoAjuste() throws SQLException {
         String sql = "SELECT id_ajuste FROM ajustes_stock_cabecera ORDER BY id_ajuste DESC LIMIT 1";
-        
-        try (PreparedStatement ps = conexion.prepareStatement(sql);
-             ResultSet rs = ps.executeQuery()) {
+
+        try (PreparedStatement ps = conexion.prepareStatement(sql); ResultSet rs = ps.executeQuery()) {
             if (rs.next()) {
                 return buscarAjustePorId(rs.getInt("id_ajuste"));
             }
@@ -166,7 +165,7 @@ public class AjusteStockDAO {
     // Obtener siguiente ajuste
     public mAjusteStock obtenerSiguienteAjuste(int idActual) throws SQLException {
         String sql = "SELECT id_ajuste FROM ajustes_stock_cabecera WHERE id_ajuste > ? ORDER BY id_ajuste ASC LIMIT 1";
-        
+
         try (PreparedStatement ps = conexion.prepareStatement(sql)) {
             ps.setInt(1, idActual);
             try (ResultSet rs = ps.executeQuery()) {
@@ -181,7 +180,7 @@ public class AjusteStockDAO {
     // Obtener anterior ajuste
     public mAjusteStock obtenerAnteriorAjuste(int idActual) throws SQLException {
         String sql = "SELECT id_ajuste FROM ajustes_stock_cabecera WHERE id_ajuste < ? ORDER BY id_ajuste DESC LIMIT 1";
-        
+
         try (PreparedStatement ps = conexion.prepareStatement(sql)) {
             ps.setInt(1, idActual);
             try (ResultSet rs = ps.executeQuery()) {
@@ -196,13 +195,13 @@ public class AjusteStockDAO {
     // Actualizar cabecera de ajuste
     public void actualizarAjusteCabecera(mAjusteStock ajuste) throws SQLException {
         String sql = "UPDATE ajustes_stock_cabecera SET observaciones = ?, aprobado = ?, estado = ? WHERE id_ajuste = ?";
-        
+
         try (PreparedStatement ps = conexion.prepareStatement(sql)) {
             ps.setString(1, ajuste.getObservaciones());
             ps.setBoolean(2, ajuste.isAprobado());
             ps.setBoolean(3, ajuste.isEstado());
             ps.setInt(4, ajuste.getIdAjuste());
-            
+
             ps.executeUpdate();
         }
     }
@@ -210,7 +209,7 @@ public class AjusteStockDAO {
     // Eliminar todos los detalles de un ajuste
     public void eliminarDetallesAjuste(int idAjuste) throws SQLException {
         String sql = "DELETE FROM ajustes_stock_detalle WHERE id_ajuste = ?";
-        
+
         try (PreparedStatement ps = conexion.prepareStatement(sql)) {
             ps.setInt(1, idAjuste);
             ps.executeUpdate();
@@ -219,14 +218,14 @@ public class AjusteStockDAO {
 
     // Buscar productos por nombre o código de barras
     public List<Object[]> buscarProductos(String busqueda) throws SQLException {
-        String sql = "SELECT pc.id_producto, pc.nombre, pd.cod_barra, pd.descripcion, " +
-                    "COALESCE(s.cantidad_disponible, 0) as stock " +
-                    "FROM productos_cabecera pc " +
-                    "INNER JOIN productos_detalle pd ON pc.id_producto = pd.id_producto " +
-                    "LEFT JOIN stock s ON pd.id_producto = s.id_producto AND pd.cod_barra = s.cod_barra " +
-                    "WHERE (pc.nombre LIKE ? OR pd.descripcion LIKE ? OR pd.cod_barra LIKE ?) " +
-                    "AND pc.estado = 1 AND pd.estado = 1 " +
-                    "ORDER BY pc.nombre, pd.descripcion LIMIT 50";
+        String sql = "SELECT pc.id_producto, pc.nombre, pd.cod_barra, pd.descripcion, "
+                + "COALESCE(s.cantidad_disponible, 0) as stock "
+                + "FROM productos_cabecera pc "
+                + "INNER JOIN productos_detalle pd ON pc.id_producto = pd.id_producto "
+                + "LEFT JOIN stock s ON pd.id_producto = s.id_producto AND pd.cod_barra = s.cod_barra "
+                + "WHERE (pc.nombre LIKE ? OR pd.descripcion LIKE ? OR pd.cod_barra LIKE ?) "
+                + "AND pc.estado = 1 AND pd.estado = 1 "
+                + "ORDER BY pc.nombre, pd.descripcion LIMIT 50";
 
         List<Object[]> productos = new ArrayList<>();
         String patron = "%" + busqueda + "%";
@@ -239,11 +238,11 @@ public class AjusteStockDAO {
             try (ResultSet rs = ps.executeQuery()) {
                 while (rs.next()) {
                     productos.add(new Object[]{
-                        rs.getInt("id_producto"),      // 0
-                        rs.getString("nombre"),        // 1
-                        rs.getString("cod_barra"),     // 2
-                        rs.getString("descripcion"),   // 3
-                        rs.getDouble("stock")          // 4
+                        rs.getInt("id_producto"), // 0
+                        rs.getString("nombre"), // 1
+                        rs.getString("cod_barra"), // 2
+                        rs.getString("descripcion"), // 3
+                        rs.getDouble("stock") // 4
                     });
                 }
             }
